Change the trivial "ed" into a more meaningful function, which calculates some kind of shortest path.
Ideally it should be defined with option monad, but it would be too complicated for now.
Even with case expression, it requires an additional consistent rule for the lifted case_option expression.

The expression is written in pure function application form rather than in case syntax for the sake of
structure consistency between ed and ed', so it may look strange now. Ultimately, the strange-looking
expressions as well as its consistency rule should all be generated by ML code.